% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QS-class.R
\docType{class}
\name{QS-class}
\alias{QS-class}
\alias{show,QS-method}
\title{Quota Share Treaty}
\usage{
\S4method{show}{QS}(object)
}
\arguments{
\item{object}{the object to display}
}
\description{
An S4 class to represent a Quota Share Treaty
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Quota Share Treaty: show method
}}

\section{Slots}{

\describe{
\item{\code{csn_clm}}{Numeric. Cession rate on claims. Should be between 0 and 1.}

\item{\code{csn_prm}}{Numeric. Cession rate on premiums. Should be between 0 and 1.}

\item{\code{com}}{Numeric. Commission rate, applied on the part of premiums given to the reinsurer. Should be between 0 and 1.}

\item{\code{ptf}}{Vector. List of portfolios on which the treaty is to be applied on.}

\item{\code{trt}}{Character. Always equal to "QS". Identifier for the type of treaty.}
}}

